security-review/requirements.txt
gitpython>=3.1.0
pydantic>=2.0.0
httpx>=0.26.0
